% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_moments_from_sens_spec}
\alias{binary_moments_from_sens_spec}
\title{Moments for binary outcome/prediction pairs from sensitivity/specificity}
\usage{
binary_moments_from_sens_spec(p, sens, spec)
}
\arguments{
\item{p}{Prevalence \eqn{P(Y = 1)}.}

\item{sens}{Classifier sensitivity \eqn{P(\hat Y = 1 \mid Y = 1)}.}

\item{spec}{Classifier specificity \eqn{P(\hat Y = 0 \mid Y = 0)}.}
}
\value{
List with \code{sigma_y2}, \code{sigma_f2}, \code{cov_y_f}, and \code{p_hat}.
}
\description{
Moments for binary outcome/prediction pairs from sensitivity/specificity
}
