% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppower-package.R
\docType{package}
\name{pppower-package}
\alias{pppower}
\alias{pppower-package}
\title{pppower: Power and Sample Size for Prediction-Powered Inference}
\description{
Tools to compute power, required sample sizes, and perform simulation
validation for Prediction-Powered Inference (PPI and PPI++).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yiqunchen.github.io/pppower}
}

}
\author{
\strong{Maintainer}: Yiqun Chen \email{yiqun.t.chen@gmail.com}

Authors:
\itemize{
  \item Moran Guo \email{mguo31@jh.edu}
}

}
\keyword{internal}
