% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_one_draw}
\alias{simulate_one_draw}
\title{Generate One Labeled + Unlabeled Sample}
\usage{
simulate_one_draw(
  n_labeled,
  n_unlabeled,
  X_sampler_L,
  X_sampler_U = NULL,
  f_generator,
  eps_sampler = function(n) rnorm(n, 0, 1),
  delta = 0
)
}
\arguments{
\item{n_labeled}{Number of labeled samples}

\item{n_unlabeled}{Number of unlabeled samples}

\item{X_sampler_L}{Function sampling labeled covariates}

\item{X_sampler_U}{Function sampling unlabeled covariates (defaults to X_sampler_L)}

\item{f_generator}{Function generating f(X)}

\item{eps_sampler}{Error sampler, default N(0,1)}

\item{delta}{Mean shift applied to f}
}
\value{
A list with labeled and unlabeled data and true mean.
}
\description{
Generate One Labeled + Unlabeled Sample
}
\keyword{internal}
